requirements.txt
djangorestframework
pyjwt==1.7.1
